Refactor Uniswap providers in MultiSwap module
