+++ UnstoppableWallet/UnstoppableWallet/Modules/MultiSwap/Providers/BaseUniswapV3MultiSwapProvider.swift
@@ -1,8 +1,6 @@
-import BigInt
 import EvmKit
 import Foundation
 import MarketKit
-import SwiftUI
 import UniswapKit
 
 class BaseUniswapV3MultiSwapProvider: BaseUniswapMultiSwapProvider {
@@ -14,7 +12,7 @@ class BaseUniswapV3MultiSwapProvider: BaseUniswapMultiSwapProvider {
         super.init(storage: storage)
     }
 
-    private func kitToken(chain: Chain, token: MarketKit.Token) throws -> UniswapKit.Token {
+    override func kitToken(chain: Chain, token: MarketKit.Token) throws -> UniswapKit.Token {
         switch token.type {
         case .native: return try kit.etherToken(chain: chain)
         case let .eip20(address): return try kit.token(contractAddress: EvmKit.Address(hex: address), decimals: token.decimals)
@@ -22,68 +20,21 @@ class BaseUniswapV3MultiSwapProvider: BaseUniswapMultiSwapProvider {
         }
     }
 
+    override func trade(rpcSource: RpcSource, chain: Chain, tokenIn: UniswapKit.Token, tokenOut: UniswapKit.Token, amountIn: Decimal, tradeOptions: TradeOptions) async throws -> Quote.Trade {
+        let bestTrade = try await kit.bestTradeExactIn(rpcSource: rpcSource, chain: chain, tokenIn: tokenIn, tokenOut: tokenOut, amountIn: amountIn, options: tradeOptions)
+        return .v3(bestTrade: bestTrade)
+    }
+
+    override func transactionData(receiveAddress: EvmKit.Address, chain: Chain, trade: Quote.Trade, tradeOptions: TradeOptions) throws -> TransactionData {
+        guard case let .v3(bestTrade) = trade else {
+            throw SwapError.invalidTrade
+        }
+
+        return try kit.transactionData(receiveAddress: receiveAddress, chain: chain, bestTrade: bestTrade, tradeOptions: tradeOptions)
+    }
+
     override func spenderAddress(chain: Chain) throws -> EvmKit.Address {
         kit.routerAddress(chain: chain)
-    }
-
-    func quote(tokenIn: MarketKit.Token, tokenOut: MarketKit.Token, amountIn: Decimal, transactionSettings: MultiSwapTransactionSettings?) async throws -> IMultiSwapQuote {
-        let blockchainType = tokenIn.blockchainType
-        let chain = evmBlockchainManager.chain(blockchainType: blockchainType)
-
-        let kitTokenIn = try kitToken(chain: chain, token: tokenIn)
-        let kitTokenOut = try kitToken(chain: chain, token: tokenOut)
-
-        guard let rpcSource = evmSyncSourceManager.httpSyncSource(blockchainType: blockchainType)?.rpcSource else {
-            throw SwapError.noHttpRpcSource
-        }
-
-        let recipient: Address? = storage.value(for: MultiSwapSettingStorage.LegacySetting.address)
-        let slippage: Decimal = storage.value(for: MultiSwapSettingStorage.LegacySetting.slippage) ?? MultiSwapSlippage.default
-
-        let kitRecipient = try recipient.map { try EvmKit.Address(hex: $0.raw) }
-
-        let tradeOptions = TradeOptions(
-            allowedSlippage: slippage,
-            ttl: TradeOptions.defaultTtl,
-            recipient: kitRecipient,
-            feeOnTransfer: false
-        )
-
-        let bestTrade = try await kit.bestTradeExactIn(rpcSource: rpcSource, chain: chain, tokenIn: kitTokenIn, tokenOut: kitTokenOut, amountIn: amountIn, options: tradeOptions)
-
-        var estimatedGas: Int?
-
-        if let evmKit = App.shared.evmBlockchainManager.evmKitManager(blockchainType: blockchainType).evmKitWrapper?.evmKit,
-           let transactionSettings, case let .evm(gasPrice, _) = transactionSettings
-        {
-            do {
-                let transactionData = try kit.transactionData(receiveAddress: evmKit.receiveAddress, chain: chain, bestTrade: bestTrade, tradeOptions: tradeOptions)
-                estimatedGas = try await evmKit.fetchEstimateGas(transactionData: transactionData, gasPrice: gasPrice)
-            } catch {}
-        }
-
-        return await Quote(
-            bestTrade: bestTrade,
-            recipient: recipient,
-            slippage: slippage,
-            estimatedGas: estimatedGas,
-            allowanceState: allowanceState(token: tokenIn, amount: amountIn)
-        )
-    }
-
-    func settingsView(tokenIn: MarketKit.Token, tokenOut _: MarketKit.Token) -> AnyView {
-        let addressViewModel = AddressMultiSwapSettingsViewModel(storage: storage, blockchainType: tokenIn.blockchainType)
-        let slippageViewModel = SlippageMultiSwapSettingsViewModel(storage: storage)
-        let viewModel = BaseMultiSwapSettingsViewModel(fields: [addressViewModel, slippageViewModel])
-        let view = ThemeNavigationView {
-            RecipientAndSlippageMultiSwapSettingsView(
-                viewModel: viewModel,
-                addressViewModel: addressViewModel,
-                slippageViewModel: slippageViewModel
-            )
-        }
-
-        return AnyView(view)
     }
 
     func swap(quote: IMultiSwapQuote, transactionSettings: MultiSwapTransactionSettings?) async throws {
@@ -95,21 +46,5 @@ class BaseUniswapV3MultiSwapProvider: BaseUniswapMultiSwapProvider {
         print(String(describing: transactionSettings))
 
         try await Task.sleep(nanoseconds: 3_000_000_000)
-    }
-}
-
-extension BaseUniswapV3MultiSwapProvider {
-    class Quote: BaseUniswapMultiSwapProvider.Quote {
-        private let bestTrade: TradeDataV3
-
-        init(bestTrade: TradeDataV3, recipient: Address?, slippage: Decimal, estimatedGas: Int?, allowanceState: AllowanceState) {
-            self.bestTrade = bestTrade
-
-            super.init(recipient: recipient, slippage: slippage, estimatedGas: estimatedGas, allowanceState: allowanceState)
-        }
-
-        override var amountOut: Decimal {
-            bestTrade.amountOut ?? 0
-        }
     }
 }
